Module validate::scatter

A Validator that creates a scatter plot from two datasets, indicating the R-value and the linear equation fitting

Example

use validate::{Validator, ScatterValidator};

let expected = vec![1., 2., 3.];
let found = vec![2., 4., 6.];

let mut validator = Validator::new("Validate Scatter", "report.html");    
// Note that we are not defining a maximum allowed error
let v = validate::ScatterValidator {
    units: Some("C"),
    expected,
    found,
    ..ScatterValidator::default()
};
validator.push(Box::new(v));

// This will not fail because we did not set a maximum allowed
// Root Mean Square Error or Mean Bias Error... if we did, it would return
// an error and the unwrap woul
validator.validate().unwrap();

Structs

ScatterValidator

Validates a time series based on Mean Bias Error and Root Mean Squared Error