Struct validate::SeriesValidator

pub struct SeriesValidator {
    pub allowed_mean_bias_error: Option<f64>,
    pub allowed_root_mean_squared_error: Option<f64>,
    pub y_units: Option<&'static str>,
    pub y_label: Option<&'static str>,
    pub x_units: Option<&'static str>,
    pub x_label: Option<&'static str>,
    pub expected_legend: Option<&'static str>,
    pub expected: Vec<f64>,
    pub found_legend: Option<&'static str>,
    pub found: Vec<f64>,
    pub chart_title: Option<&'static str>,
}

Validates a time series based on Mean Bias Error and Root Mean Squared Error

Fields
allowed_mean_bias_error: Option<f64>

The maximum allowed Mean Bias Error

allowed_root_mean_squared_error: Option<f64>

The maximum allowed Root Mean Squared Error

y_units: Option<&'static str>

The units in the y axis of the chart

y_label: Option<&'static str>

The label in the y axis of the chart

x_units: Option<&'static str>

The units in the x axis of the chart

x_label: Option<&'static str>

The label in the x axis of the chart

expected_legend: Option<&'static str>

The name of the series caled expected

expected: Vec<f64>

The time series containing the expected values

found_legend: Option<&'static str>

The name of the found time series

found: Vec<f64>

The time series containing the found values

chart_title: Option<&'static str>

the title of the chart