Module validate::time_series

A Validator that plots two time series and calculates—if required—the Root Mean Squared Error and Mean Bias Error between them.

Example

use validate::{Validator, SeriesValidator};

let expected = vec![1., 2., 3.];
let found = vec![5., 6., 6.];

let mut validator = Validator::new("Validate Time series", "report.html");    
// Note that we are not defining a maximum allowed error
let v = validate::SeriesValidator {
    x_label: Some("time step"),
    y_label: Some("Zone Temperature"),
    y_units: Some("C"),
    expected,
    found,
    ..SeriesValidator::default()
};
validator.push(Box::new(v));

// This will not fail because we did not set a maximum allowed
// Root Mean Square Error or Mean Bias Error... if we did, it would return
// an error and the unwrap woul
validator.validate().unwrap();

Structs

SeriesValidator

Validates a time series based on Mean Bias Error and Root Mean Squared Error