Trait validate::Validate

pub trait Validate {
    fn validate(&self) -> ValidationResult;
}

The main trait of this crate. All validator modules need to comply with this trait.

Required Methods

fn validate(&self) -> ValidationResult

Runs a validation procedure, returning an error message if the validation failed.

Whether it fails or not, it should write the results of the validations into file so a full report is written.