Crate validate

This crate was developed with the intent of helping you to validate scientific tools; for exmple, for comparing the results of the temperature calculated by an algorithm and those measured in an experiment. It is supposed to be embedded the unit or integration testing.

Different Validation tasks can be created by implementing the Validate trait.
These tasks will be stored into a Validator object which will write a report into an HTML File.

Example

use validate::{valid, Validator, Validate, ValidationResult};

// This checks that two numbers are equal
struct CustomValidator {
    expected: u8,
    found: u8,
    title: &'static str,
}

impl Validate for CustomValidator {
    fn validate(&self)->ValidationResult{
        let valid = self.expected == self.found;
        let mut ret = String::new();
        let mut err = String::new();
         
         
        // We return an error if the validation is not succesful, but we still
        // write something into the report. Even if this particular Validation
        // fails, the Validations object will run all the validations and print
        // the error messages into the STDERR
        if valid {
            ret = format!("## {}\n\n * Passed! {} and {} are equal", self.title, self.expected, self.found);
        }else{
            err = format!("{}\n{} and {} aren't equal", err, self.expected, self.found);
            ret = format!("{}\n\n# {}\n\n * Failed... {} and {} aren't equal",ret, self.title, self.expected, self.found );
        };
        if err.len()>0{
            ValidationResult::Err(err, ret)
        }else{
            ValidationResult::Ok(ret)
        }
    }
}

/// Some explanation about the validation
///
/// It is always important to know what is it that we are validating
#[valid(Some Validation)]
fn check_if_equal()->Box<dyn Validate>{
    let v = CustomValidator{
        expected: 2,
        found: 2,
        title: "Check that 2 and 3 are equal"
    };
    Box::new(v) // We need a Box.
}

// Write a test
#[test]
fn test_custon_validator(){
    let mut validator = Validator::new("Test Validation", "report.html");
    validator.push(check_if_equal());
    validator.validate().unwrap()
}

Modules

numberish 🔒

A trait defining some numerical-ish trait.

scatter 🔒

A Validator that creates a scatter plot from two datasets, indicating the R-value and the linear equation fitting

stats

Module with some useful functions for calculating indicators for validation (e.g., Mean Squared Error)

time_series 🔒

A Validator that plots two time series and calculates—if required—the Root Mean Squared Error and Mean Bias Error between them.

validator_wrapper 🔒

A wrapper that contains an object that implements Validate

Macros

assert_close

Asserts whether two numbers are close enough by comparing the first argument with the second, and the threshold being the third.

assert_not_close

Asserts whether two numbers are close enough

Structs

ScatterValidator

Validates a time series based on Mean Bias Error and Root Mean Squared Error

SeriesValidator

Validates a time series based on Mean Bias Error and Root Mean Squared Error

Validator

This structure holds a number of validations to be ran, runs them, and writes the results into an HTML report. It has a title, which is used as a Header in its report.

ValidatorWrapper

A wrapper that contains an object that implements Validate

Enums

ValidationResult

Implements a validation error, where Ok returns just the text to write in the report, but Err returns not only that but also an error message

Traits

Validate

The main trait of this crate. All validator modules need to comply with this trait.

Functions

from_csv

Reads a number of columns from a CSV, transforms them into f64

Attribute Macros

valid

An Attribute MAcro