Module validate::stats

Module with some useful functions for calculating indicators for validation (e.g., Mean Squared Error)

Functions

linear_coefficients

Calculates the coefficients $a$ and $b$ that best fit the model $y = a + b\times x$. Returns $a$, $b$ and the $R^2$ of the fit.

mean

Calculates the mean of a dataset

mean_bias_error

Calculates the Mean Bias Error between to datasets, indicating whether one of them is consistently larger or smaller than the other.

min_max

Calculates the maximum and minimum in a series.

root_mean_squared_error

Calculates the Root Mean Squared Error between to datasets, indicating the average absolute difference between them.